feat: Migrate from CRA to vite
This project was created using create-react-app. Since this is method
was officially deprecated we need to migrate to a more recent method
to build and develop the app.
To ease the migration, we are gonna use Vite instead of a more complex
framework like Next.js.

This commit installs the needed libraries and adds the proper
configuration to run and build the project using Vite.