fix: Attempt to fix google login
The google login button was not working since the library we were
using was deprecated. We switched to a new library but there is an
issue after login in, the popup cannot talk with the app.

This commit adds those changes to keep working on a solution to
fix it.
If that doesn't work, we will migrate to an IdP like Auth0.